requirements.txt
python 2.7.12
nflgame-redux
virtualenv 16.0.0